A critical bug in beam element has been corrected
New beam element based on the Timoshenko theory:
- the main "beam" is modified to get a shear capability. For that purpose, use the function
fedoo.Element.SetProperties_Beam(Iyy, Izz, A, nu=None, k=1, E= None, G=None) before creating the assembly
If the properties are not defined, the bernouilli Beam is used.
In this element, the shape function use the beam properties, therefore this is not compatible with the ParametricBeam weakform
- a new "beamFCQ" element is also added which is compatible with the ParametricBeam weakform, but has the drawback to require no physical internal dof
To add internal nodes, a method AddInternalNodes(numberOfInternalNodes) has been added to the mesh object.

A unique "WeakForm_Beam" weakform is used for the 2 beam elements and can account for shear effect or not (default behavior)

A new simple example for beam element has been added.

And perhaps other minor changes